backend/requirements.txt
django==4.1.7
stripe==5.1.1
psycopg2==2.9.5
python-dotenv==0.21.1
types-requests==2.28.11.*